Extract npm-specific logic to src/npm.ts
